- fix logical children
